requirements.txt
python-telegram-bot==20.3
requests==2.31.0
gspread==5.11.2
oauth2client==4.1.3
python-dotenv==1.0.1